add pandas extra and enable etl pandas tests
